Rename evaluation api
(cherry picked from commit 96416b318425ea4e5077e06d6055ea6c4eca9e70)
