+++ packages/frontend/editor-ui/src/components/Evaluations/EditDefinition/sections/RunsSection.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import type { TestRunRecord } from '@/api/testDefinition.ee';
+import type { TestRunRecord } from '@/api/evaluation.ee';
 import MetricsChart from '@/components/Evaluations/ListRuns/MetricsChart.vue';
 import TestRunsTable from '@/components/Evaluations/ListRuns/TestRunsTable.vue';
 import { useI18n } from '@/composables/useI18n';
