+++ packages/frontend/editor-ui/src/components/Evaluations/ListRuns/TestRunsTable.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import type { TestRunRecord } from '@/api/testDefinition.ee';
+import type { TestRunRecord } from '@/api/evaluation.ee';
 import { useI18n } from '@/composables/useI18n';
 import { N8nIcon, N8nText } from '@n8n/design-system';
 import type { IconColor } from '@n8n/design-system/types/icon';
