+++ packages/frontend/editor-ui/src/views/Evaluations/TestRunDetailView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import type { TestCaseExecutionRecord } from '@/api/testDefinition.ee';
+import type { TestCaseExecutionRecord } from '@/api/evaluation.ee';
 import type { TestTableColumn } from '@/components/Evaluations/shared/TestTableBase.vue';
 import TestTableBase from '@/components/Evaluations/shared/TestTableBase.vue';
 import { useI18n } from '@/composables/useI18n';
